+++ problem/edge_cover.go
@@ -1,8 +1,6 @@
 package problem
 
 import (
-	"github.com/roidaradal/fn/dict"
-	"github.com/roidaradal/fn/list"
 	"github.com/roidaradal/opt/data"
 	"github.com/roidaradal/opt/discrete"
 	"github.com/roidaradal/opt/fn"
@@ -26,15 +24,6 @@ func edgeCover(name string) *discrete.Problem {
 		return nil
 	}
 
-	p.AddUniversalConstraint(func(solution *discrete.Solution) bool {
-		// Check for all vertices, covered by at least one edge endpoint in solution subset
-		count := dict.NewCounter(graph.Vertices)
-		for _, x := range fn.AsSubset(solution) {
-			v1, v2 := graph.Edges[x].Tuple()
-			count[v1] += 1
-			count[v2] += 1
-		}
-		return list.AllGreater(dict.Values(count), 0)
-	})
+	p.AddUniversalConstraint(fn.ConstraintAllVerticesCovered(graph.Graph, graph.Vertices))
 	return p
 }